+++ test/test.config.js
@@ -1,0 +1,7 @@
+const chai = require('chai');
+let sinonChai = require('sinon-chai');
+chai.use(require('chai-as-promised'));
+chai.use(sinonChai);
+
+global.expect = chai.expect;
+global.sinon = require('sinon');
